Fix memory error
We allocate TileInfo objects on the heap. We used a
std.ArrayList(TileInfo) during buildFillingInfo, but of course the
memory address of each element becomes invalid when the list has to
reallocate to expand.

With this commit we calculate the necessary number of TileInfo objects
up front and use a slice, which does not reallocate and hence has memory
addresses that remain valid.